Add new rail timeslider demo
Need to figure out the proper way to read the rail.json file into the demo directly. Currently there's a copy/paste into a JS file. I tried variants of the fetch API, but couldn't figure that out quickly.

This is to help with https://github.com/OpenHistoricalMap/issues/issues/405